Move types to definitions
